services/python-app/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-multipart==0.0.20
psycopg2-binary==2.9.9
SQLAlchemy>=1.4,<2.0
pandas==2.1.4
py-postgresql==1.3.0
PyHDFS==0.3.1
PyHive==0.7.0
thrift==0.20.0
thrift-sasl==0.4.3
clickhouse-driver==0.2.8
pyarrow==17.0.0